+++ src/time/mod.rs
@@ -16,6 +16,8 @@ use serde::{Deserialize, Serialize};
 use std::error::Error;
 #[cfg(feature = "std")]
 use std::time::{SystemTime, SystemTimeError};
+#[cfg(feature = "std")]
+pub use std_mod::*;
 
 pub mod ascii;
 pub mod cds;
@@ -64,45 +66,11 @@ pub enum TimestampError {
     /// Contains tuple where first value is the expected time code and the second
     /// value is the found raw value
     InvalidTimeCode(CcsdsTimeCodes, u8),
-    ByteConversionError(ByteConversionError),
-    CdsError(cds::CdsError),
-    CucError(cuc::CucError),
+    ByteConversion(ByteConversionError),
+    Cds(cds::CdsError),
+    Cuc(cuc::CucError),
     DateBeforeCcsdsEpoch(DateTime<Utc>),
     CustomEpochNotSupported,
-}
-
-impl From<cds::CdsError> for TimestampError {
-    fn from(e: cds::CdsError) -> Self {
-        TimestampError::CdsError(e)
-    }
-}
-
-impl From<cuc::CucError> for TimestampError {
-    fn from(e: cuc::CucError) -> Self {
-        TimestampError::CucError(e)
-    }
-}
-
-#[cfg(feature = "std")]
-#[cfg_attr(doc_cfg, doc(cfg(feature = "std")))]
-#[derive(Debug, Clone)]
-pub enum StdTimestampError {
-    SystemTimeError(SystemTimeError),
-    TimestampError(TimestampError),
-}
-
-#[cfg(feature = "std")]
-impl From<TimestampError> for StdTimestampError {
-    fn from(v: TimestampError) -> Self {
-        Self::TimestampError(v)
-    }
-}
-
-#[cfg(feature = "std")]
-impl From<SystemTimeError> for StdTimestampError {
-    fn from(v: SystemTimeError) -> Self {
-        Self::SystemTimeError(v)
-    }
 }
 
 impl Display for TimestampError {
@@ -114,13 +82,13 @@ impl Display for TimestampError {
                     "invalid raw time code value {raw_val} for time code {time_code:?}"
                 )
             }
-            TimestampError::CdsError(e) => {
+            TimestampError::Cds(e) => {
                 write!(f, "cds error {e}")
             }
-            TimestampError::CucError(e) => {
+            TimestampError::Cuc(e) => {
                 write!(f, "cuc error {e}")
             }
-            TimestampError::ByteConversionError(e) => {
+            TimestampError::ByteConversion(e) => {
                 write!(f, "byte conversion error {e}")
             }
             TimestampError::DateBeforeCcsdsEpoch(e) => {
@@ -137,11 +105,38 @@ impl Display for TimestampError {
 impl Error for TimestampError {
     fn source(&self) -> Option<&(dyn Error + 'static)> {
         match self {
-            TimestampError::ByteConversionError(e) => Some(e),
-            TimestampError::CdsError(e) => Some(e),
-            TimestampError::CucError(e) => Some(e),
+            TimestampError::ByteConversion(e) => Some(e),
+            TimestampError::Cds(e) => Some(e),
+            TimestampError::Cuc(e) => Some(e),
             _ => None,
         }
+    }
+}
+impl From<cds::CdsError> for TimestampError {
+    fn from(e: cds::CdsError) -> Self {
+        TimestampError::Cds(e)
+    }
+}
+
+impl From<cuc::CucError> for TimestampError {
+    fn from(e: cuc::CucError) -> Self {
+        TimestampError::Cuc(e)
+    }
+}
+
+#[cfg(feature = "std")]
+#[cfg_attr(doc_cfg, doc(cfg(feature = "std")))]
+pub mod std_mod {
+    use crate::time::TimestampError;
+    use std::time::SystemTimeError;
+    use thiserror::Error;
+
+    #[derive(Debug, Clone, Error)]
+    pub enum StdTimestampError {
+        #[error("system time error: {0}")]
+        SystemTime(#[from] SystemTimeError),
+        #[error("timestamp error: {0}")]
+        Timestamp(#[from] TimestampError),
     }
 }